requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-shortuuidfield==0.1.3
djangorestframework==3.15.2
et_xmlfile==2.0.0
gevent==24.11.1
greenlet==3.1.1
idna==3.10
kombu==5.4.2
mysqlclient==2.2.7
numpy==2.2.3
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.50
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.1
redis==5.2.1
requests==2.32.3
setuptools==75.8.2
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==7.2
